refactor: VOMS class does not inherit from BaseSecurity
